+++ src/main/java/com/bootdo/common/controller/LogController.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.aspectj.weaver.tools.cache.AsynchronousFileCacheBacking.RemoveCommand;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -22,18 +23,20 @@ import com.bootdo.common.utils.R;
 
 @RequestMapping("/common/log")
 @Controller
-public class LogController {
+public class LogController extends BaseController{
 	@Autowired
 	LogService logService;
 	String prefix = "common/log";
 
 	@GetMapping()
+	@RequiresPermissions("sys:log:lists")
 	String log() {
 		return prefix + "/log";
 	}
 
 	@ResponseBody
 	@GetMapping("/list")
+	@RequiresPermissions("sys:log:list")
 	PageDO<LogDO> list(@RequestParam Map<String, Object> params) {
 		Query query = new Query(params);
 		PageDO<LogDO> page = logService.queryList(query);
@@ -42,7 +45,11 @@ public class LogController {
 	
 	@ResponseBody
 	@PostMapping("/remove")
+	@RequiresPermissions("sys:log:remove")
 	R remove(Long id) {
+		if ("test".equals(getUsername())) {
+			return R.error(1, "演示系统不允许修改,完整体验请部署程序");
+		}
 		if (logService.remove(id)>0) {
 			return R.ok();
 		}
@@ -51,7 +58,11 @@ public class LogController {
 
 	@ResponseBody
 	@PostMapping("/batchRemove")
+	@RequiresPermissions("sys:log:batchRemove")
 	R batchRemove(@RequestParam("ids[]") Long[] ids) {
+		if ("test".equals(getUsername())) {
+			return R.error(1, "演示系统不允许修改,完整体验请部署程序");
+		}
 		int r = logService.batchRemove(ids);
 		if (r > 0) {
 			return R.ok();
